Move components into folders, and simplify imports
