+++ src/index.tsx
@@ -1,9 +1,8 @@
 import React from "react"
 
+import App from "pages/App"
 import ReactDOM from "react-dom/client"
-
-import App from "./App"
-import reportWebVitals from "./reportWebVitals"
+import reportWebVitals from "reportWebVitals"
 
 import "./index.css"
