+++ src/pages/App.tsx
@@ -7,11 +7,12 @@ import "ag-grid-enterprise"
 import "ag-grid-community/styles/ag-grid.css"
 import "ag-grid-community/styles/ag-theme-alpine.css"
 
+import NavBar from "components/NavBar"
+import FakeGetJobsService from "services/mocks/FakeGetJobsService"
+import FakeGroupJobsService from "services/mocks/FakeGroupJobsService"
+import { makeTestJobs } from "utils"
+
 import JobsPage from "./JobsPage"
-import NavBar from "./NavBar"
-import FakeGetJobsService from "./services/mocks/FakeGetJobsService"
-import FakeGroupJobsService from "./services/mocks/FakeGroupJobsService"
-import { makeTestJobs } from "./util"
 
 import "./App.css"
