+++ src/services/mocks/FakeGetJobsService.ts
@@ -1,6 +1,6 @@
-import { Job, JobFilter, JobKey, JobOrder } from "../../model"
-import { compareValues, mergeFilters } from "../../util"
-import { GetJobsService } from "../GetJobsService"
+import { Job, JobFilter, JobKey, JobOrder } from "model"
+import { GetJobsService } from "services/GetJobsService"
+import { compareValues, mergeFilters } from "utils"
 
 export default class FakeGetJobsService implements GetJobsService {
   jobs: Job[]
