+++ src/services/mocks/FakeGroupJobsService.ts
@@ -1,6 +1,6 @@
-import { Job, JobFilter, JobGroup, JobKey, JobOrder } from "../../model"
-import { compareValues, mergeFilters } from "../../util"
-import { GroupJobsService } from "../GroupJobsService"
+import { Job, JobFilter, JobGroup, JobKey, JobOrder } from "model"
+import { GroupJobsService } from "services/GroupJobsService"
+import { compareValues, mergeFilters } from "utils"
 
 export default class FakeGroupJobsService implements GroupJobsService {
   jobs: Job[]
